refactor: delete inedx files
