--- src/data/protocols/http/index.ts
@@ -1,2 +1,0 @@
-export * from './http-get-client';
-export * from './http-response';
